# build/build/swagger.json
{
    "openapi": "3.0.0",
    "components": {
        "examples": {},
        "headers": {},
        "parameters": {},
        "requestBodies": {},
        "responses": {},
        "schemas": {
            "registerRole": {
                "type": "string",
                "enum": [
                    "seller",
                    "customer"
                ]
            },
            "role": {
                "anyOf": [
                    {
                        "$ref": "#/components/schemas/registerRole"
                    },
                    {
                        "type": "string",
                        "enum": [
                            "admin"
                        ]
                    }
                ]
            },
            "ICreateUser": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "$ref": "#/components/schemas/role"
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "username",
                    "password",
                    "email",
                    "role"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "mongoose.Types.ObjectId": {
                "type": "string"
            },
            "IProduct": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "double"
                    },
                    "quantity": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    },
                    "seller_id": {
                        "$ref": "#/components/schemas/mongoose.Types.ObjectId"
                    },
                    "deactivated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "name",
                    "price",
                    "quantity",
                    "seller_id"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "ILoginUser": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "username",
                    "password"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "IRegisterUser": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "role": {
                        "$ref": "#/components/schemas/registerRole"
                    },
                    "image": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "username",
                    "password",
                    "email",
                    "role"
                ],
                "type": "object",
                "additionalProperties": false
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "td_ecommerce",
        "version": "1.0.0",
        "description": "This project is part of the **Talentos Digitales Bootcamp** and focuses on developing a full-stack e-commerce website from scratch. The goal is to build a fully functional e-commerce platform, applying skills in both backend and frontend development.",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "paths": {
        "/api/users": {
            "post": {
                "operationId": "CreateOneCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ICreateUser"
                            }
                        }
                    }
                }
            },
            "get": {
                "operationId": "GetAllActiveCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/api/users/deleted": {
            "get": {
                "operationId": "GetAllDeletedCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/api/users/include_deleted": {
            "get": {
                "operationId": "GetAllCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/api/users/{id}": {
            "get": {
                "operationId": "GetOneCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateOneCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ICreateUser"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteOneCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/users/hard/{id}": {
            "delete": {
                "operationId": "DeleteOneHardCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/products": {
            "post": {
                "operationId": "CreateOneCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IProduct"
                            }
                        }
                    }
                }
            },
            "get": {
                "operationId": "GetAllActiveCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/api/products/deleted": {
            "get": {
                "operationId": "GetAllDeletedCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/api/products/include_deleted": {
            "get": {
                "operationId": "GetAllCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [],
                "parameters": []
            }
        },
        "/api/products/{id}": {
            "get": {
                "operationId": "GetOneCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            },
            "put": {
                "operationId": "UpdateOneCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IProduct"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteOneCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/products/hard/{id}": {
            "delete": {
                "operationId": "DeleteOneHardCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Products"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "operationId": "LoginCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ILoginUser"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "operationId": "RegisterCtrl",
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "tags": [
                    "Auth"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IRegisterUser"
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}